requirements.txt
Flask==3.1.2
Werkzeug==3.1.3
requests==2.31.0
python-dotenv==1.0.0
gTTS==2.3.2
soundfile==0.12.1
numpy==1.25.2
pydub==0.25.1
gunicorn==21.2.0
